linters/taskfile-lint/pyproject.toml
[project]
name = "taskfile-lint"
version = "0.1.0"
description = "Linter for taskfiles"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.2",
]

[project.scripts]
taskfile-lint = "taskfile_lint.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "ruff>=0.9.3",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
isort.order-by-type = false
select = ["E", "I", "F"]
